api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-multipart==0.0.6
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
